requirements.txt

langchain>=0.0.383
openai>=0.27.0
python-dotenv>=1.0.0
